Initial stages of new render queue; extended cbuffer map/unmap support to all buffer types

+++ RJ/GameVarsExtern.cpp
@@ -131,7 +131,7 @@ namespace Game {
 	// Rendering constants
 	const bool C_RENDER_DEBUG_LAYER = false;				// Flag indicating whether we should attempt to load a debug layer for the current rendering 
 															// engine (e.g D3D_DEVICE_DEBUG). Only available in debug builds regardless of the state of this flag
-	const unsigned int C_INSTANCED_RENDER_LIMIT = 1000U;	// The maximum number of instances that can be rendered in any one draw call by the engine
+	const size_t C_INSTANCED_RENDER_LIMIT = 1000U;			// The maximum number of instances that can be rendered in any one draw call by the engine
 	const float C_MODEL_SIZE_LIMIT = 10000.0f;				// The maximum size of any model; prevents overflow / accidental scaling to unreasonble values
 	const int C_MAX_ARTICULATED_MODEL_SIZE = 128;			// The maximum number of components within any articulated model
 	const unsigned int C_DEFAULT_RENDERQUEUE_CHECK_INTERVAL = 1000U;		// Time (ms) between pre-optimisation checks of the render queue
